refactor(backend): small refactoring

+++ apps/backend/src/services/auth.service.ts
@@ -116,6 +116,7 @@ export class AuthService {
     }
 
     async logout(refreshToken: string) {
+        if (!refreshToken) return;
         return tokenService.removeToken(refreshToken);
     }
 }
